% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biips_model.r
\name{biips_model}
\alias{biips_model}
\title{Create a biips model object described in BUGS language.}
\usage{
biips_model(file, data = parent.frame(), sample_data = TRUE,
  quiet = FALSE)
}
\arguments{
\item{file}{string. path of the BUGS file which defines the stochastic
model. Alternatively, \code{file} can be a readable text-mode connection,
or a complete URL.}

\item{data}{either a list or environment containing constants and
observed values or a character vector containing names of variables
existing in the workspace. Any numeric objects in \code{data} corresponding
to node arrays used in \code{file} are taken to represent the values of
observed nodes in the model.}

\item{sample_data}{logical. Toggle the evaluation of the 'data' block of
the BUGS model that generates or transforms data. (default = \code{TRUE}).}

\item{quiet}{logical. Deactivate verbosity. (default = \code{FALSE}).}
}
\value{
An object of class \code{biips} which can be used to generate samples
  from the posterior distribution of the parameters.

  An object of class \code{biips} is a list of functions that share a common
  environment. The functions can be used to query information on the model.
  \item{ptr()}{Get an external pointer to the compiled model object.}
  \item{file()}{Get a string. filename of the BUGS model.}
  \item{data()}{Get a named list of the original data of the model.}
  \item{model()}{Get a character vector. BUGS language definition of the model.}
  \item{recompile()}{Recompile the model using the original data set.}
}
\description{
Create a biips model object described in BUGS language.
}
\examples{
modelfile <- system.file('extdata', 'hmm.bug', package = 'rbiips')
stopifnot(nchar(modelfile) > 0)
cat(readLines(modelfile), sep = '\\n')

data <- list(tmax = 10, p = c(.5, .5), logtau_true = log(1), logtau = log(1))
model <- biips_model(modelfile, data, sample_data = TRUE)

\dontrun{
tmax <- 10
p <- c(.5, .5)
logtau_true <- log(1)
logtau <- logtau_true

datanames <- c('tmax', 'p', 'logtau_true', 'logtau')
model <- biips_model(modelfile, datanames, sample_data = TRUE)
}

is.biips(model)
print(model)

model$data()

variable.names(model)
biips_variable_names(model)

biips_nodes(model)

\dontrun{
dotfile <- 'hmm.dot'
biips_print_dot(model, dotfile)
cat(readLines(dotfile), sep = '\\n')
}
}
\seealso{
\code{\link{biips_variable_names}}, \code{\link{biips_nodes}},
  \code{\link{biips_print_dot}}, \code{\link{biips_add_function}},
  \code{\link{biips_add_distribution}}
}
\keyword{graphs}
\keyword{models}
