% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.r
\name{biips_smc_sensitivity}
\alias{biips_smc_sensitivity}
\title{Sensitivity analysis of parameters with SMC}
\usage{
biips_smc_sensitivity(object, param_values, n_part, ...)
}
\arguments{
\item{object}{\code{biips} model object as returned by
\code{\link{biips_model}}.}

\item{param_values}{named list of numeric values. Each member is named after
a parameter for which we want to study sensitivity. Each member value is a
vector or array whose last dimension is equal to the number of values for
which we want to study sensitivity.}

\item{n_part}{integer. Number of particles.}

\item{...}{Additional arguments to be passed to the SMC
algorithm such as \code{rs_thres} and \code{rs_type}.
See \code{\link{biips_smc_samples}}  for more details.}
}
\value{
\code{biips_smc_sensitivity} returns a list with the following
  members:
  \item{log_marg_like}{vector of log marginal likelihood estimates
    at the different values of the parameters.}
  \item{log_marg_like_pen}{vector of penalized log marginal likelihood
      estimates at the different values of the parameters.}
  \item{max_param}{parameters value with maximum \code{log_marg_like}.}
  \item{max_log_marg_like}{maximum \code{log_marg_like} value.}
  \item{max_param_pen}{parameters value with maximum \code{log_marg_like_pen}.}
  \item{max_log_marg_like_pen}{maximum \code{log_marg_like_pen} value.}
}
\description{
Sensitivity analysis of parameters with SMC
}
\examples{
modelfile <- system.file('extdata', 'hmm.bug', package = 'rbiips')
stopifnot(nchar(modelfile) > 0)
cat(readLines(modelfile), sep = '\\n')

data <- list(tmax = 10, p = c(.5, .5), logtau_true = log(1), logtau = log(1))
model <- biips_model(modelfile, data)

n_part <- 50
logtau_val <- -10:10
out_sens <- biips_smc_sensitivity(model, list(logtau = logtau_val), n_part)

par(mfrow = c(2, 1))
plot(logtau_val, out_sens$log_marg_like, type = 'l', col = 'blue',
     xlab = 'logtau', ylab = 'log p(y|logtau) ', main = 'SMC sensitivity')
points(data$logtau, min(out_sens$log_marg_like), pch = 17, col = 'green')

plot(logtau_val, out_sens$log_marg_like_pen, type = 'l', col = 'blue',
     xlab = 'logtau', ylab = 'log p(y|logtau) + log p(logtau)')
plml <- out_sens$log_marg_like_pen
ymin <- min(plml[is.finite(plml)])
points(data$logtau, ymin, pch = 17, col = 'green')
}
\seealso{
\code{\link{biips_model}}, \code{\link{biips_smc_samples}}
}
