% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biips-object.r
\name{clear_monitors.biips}
\alias{clear_monitors.biips}
\title{Clear monitors}
\usage{
\method{clear_monitors}{biips}(object, type = "fsb", release_only = FALSE,
  ...)
}
\arguments{
\item{object}{\code{biips} model object}

\item{type}{string with characters \code{'f'} (filtering), \code{'s'}
(smoothing) and/or \code{'b'} (backward smoothing).}

\item{release_only}{logical. If \code{TRUE}, only releases memory occupied by
monitors. Information about monitored nodes is still present. If
\code{FALSE} clears all information about monitored nodes as well as
memory.}
}
\description{
Clear monitors
}
\keyword{internal}

