% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{data_preprocess}
\alias{data_preprocess}
\title{Preprocess the data (list or env) to list.}
\usage{
data_preprocess(data)
}
\arguments{
\item{data}{A list or an environment of objects.}
}
\description{
Preprocess the data (list or env) to list.
}
\details{
stop if no-name lists; duplicate names. stop if the objects given
  name is not found. remove NULL, non-numeric elements
}
\author{
Adapted from \pkg{rstan}.
}
\keyword{internal}

