% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{out_pimh}
\alias{out_pimh}
\title{Sample output from \code{\link{biips_pimh_samples}}}
\format{An object of class \code{mcmcarray.list} of length 3.}
\usage{
out_pimh
}
\description{
Sample output from \code{\link{biips_pimh_samples}}
}
\keyword{datasets}

