% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{out_pmmh}
\alias{out_pmmh}
\title{Sample output from \code{\link{biips_pmmh_samples}}}
\format{An object of class \code{mcmcarray.list} of length 4.}
\usage{
out_pmmh
}
\description{
Sample output from \code{\link{biips_pmmh_samples}}
}
\keyword{datasets}
