% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{out_smc}
\alias{out_smc}
\title{Sample output from \code{\link{biips_smc_samples}}}
\format{An object of class \code{smcarray.fsb.list} of length 3.}
\usage{
out_smc
}
\description{
Sample output from \code{\link{biips_smc_samples}}
}
\keyword{datasets}
