% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{parse_varname}
\alias{parse_varname}
\title{Try to parse string of form 'a' or 'a[n,p:q,r]' where 'a' is a variable name
and n,p,q,r are integers}
\usage{
parse_varname(varname)
}
\arguments{
\item{varname}{string containing the name of the variable to sparse}
}
\description{
Try to parse string of form 'a' or 'a[n,p:q,r]' where 'a' is a variable name
and n,p,q,r are integers
}
\keyword{internal}

