% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc.r
\name{smc_forward_algo.biips}
\alias{smc_forward_algo.biips}
\title{Run a forward SMC algorithm}
\usage{
\method{smc_forward_algo}{biips}(object, n_part, rs_thres = 0.5,
  rs_type = "stratified", seed = get_seed(), ...)
}
\arguments{
\item{object}{biips model object}

\item{n_part}{integer. number of particles}

\item{rs_thres}{real. resampling threshold.}

\item{rs_type}{string. Name of the resampling algorithm}

\item{seed}{integer. RNG seed passed to Biips SMC algorithm}

\item{...}{unused}
}
\value{
A logical. \code{TRUE} if success.
}
\description{
Run a forward SMC algorithm
}
\keyword{internal}
