% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.r
\name{plot-methods}
\alias{plot-methods}
\alias{plot.density.smcarray.fsb.univariate}
\alias{plot.table.smcarray.fsb.univariate}
\alias{plot.density.smcarray}
\alias{plot.table.smcarray}
\alias{plot.density.smcarray.fsb}
\alias{plot.table.smcarray.fsb}
\alias{plot.density.smcarray.fsb.list}
\alias{plot.table.smcarray.fsb.list}
\alias{plot-methods}
\alias{plot.table.mcmcarray}
\alias{plot.density.mcmcarray.univariate}
\alias{plot.density.mcmcarray}
\alias{plot.histogram.mcmcarray}
\alias{plot.table.mcmcarray.list}
\alias{plot.density.mcmcarray.list}
\alias{plot.histogram.mcmcarray.list}
\title{Plot methods.}
\usage{
\method{plot}{density.smcarray.fsb.univariate}(x, type = "l", col = 1:6,
  pch = NULL, lwd = NULL, lty = NULL, main = NULL, xlab = NULL,
  ylab = "Density", xlim = range(unlist(lapply(x, function(x) x$x))),
  ylim = c(0, max(unlist(lapply(x, function(x) x$y)))), ...)

\method{plot}{table.smcarray.fsb.univariate}(x, type = "h", col = 1:6,
  pch = NULL, lwd = 2, lty = NULL, main = NULL, xlab = NULL,
  ylab = "Probability", xlim, ylim = c(0, max(unlist(x))), width = 0.1,
  ...)

\method{plot}{density.smcarray}(x, main = NULL, xlab = NULL,
  ylab = "Density", ...)

\method{plot}{table.smcarray}(x, main = NULL, xlab = NULL,
  ylab = "Probability", ...)

\method{plot}{density.smcarray.fsb}(x, main = NULL, xlab = NULL,
  ylab = "Density", ...)

\method{plot}{table.smcarray.fsb}(x, main = NULL, xlab = NULL,
  ylab = "Probability", ...)

\method{plot}{density.smcarray.fsb.list}(x, main = NULL, xlab = NULL,
  ylab = "Density", ...)

\method{plot}{table.smcarray.fsb.list}(x, main = NULL, xlab = NULL,
  ylab = "Probability", ...)

\method{plot}{table.mcmcarray}(x, main = NULL, xlab = NULL,
  ylab = "Probability", ...)

\method{plot}{density.mcmcarray.univariate}(x, type = "l",
  main = x$data.name, ...)

\method{plot}{density.mcmcarray}(x, main = NULL, xlab = NULL, ...)

\method{plot}{histogram.mcmcarray}(x, main = NULL, xlab = NULL, ...)

\method{plot}{table.mcmcarray.list}(x, main = NULL, xlab = NULL, ...)

\method{plot}{density.mcmcarray.list}(x, main = NULL, xlab = NULL,
  ylab = "Density", ...)

\method{plot}{histogram.mcmcarray.list}(x, main = NULL, xlab = NULL, ...)
}
\arguments{
\item{x}{a density object.}

\item{type, col, pch, lwd, lty}{plotting parameters with useful defaults.}

\item{main, xlab, ylab}{plotting parameters with useful defaults.}

\item{xlim, ylim}{plotting parameters with useful defaults.}

\item{...}{further plotting parameters. See
\code{\link[stats]{plot.density}}, \code{\link[graphics]{plot.histogram}},
\code{\link[graphics]{plot.table}}}

\item{width}{real. width of spacing for bars at the same value.}
}
\value{
\code{NULL}
}
\description{
Plot methods.
}
\examples{
#' # SMC algorithm
data(out_smc)
par(mfrow = c(2, 2))
plot(biips_density(out_smc$x, bw = 'nrd0', adjust = 1, n = 100))
plot(biips_table(out_smc[['c[2:10]']]))
plot(biips_density(out_smc$x$f))
plot(biips_table(out_smc[['c[2:10]']]$f))

#' # PIMH algorithm
data(out_pimh)
par(mfrow = c(2, 2))
plot(biips_density(out_pimh$x))
plot(biips_table(out_pimh[['c[2:10]']]))

#' # PMMH algorithm
data(out_pmmh)
plot(biips_density(out_pmmh$logtau))
par(mfrow = c(2, 2))
plot(biips_density(out_pmmh$x))
plot(biips_table(out_pmmh[['c[2:10]']]))
}
\seealso{
\code{\link[stats]{plot.density}},
  \code{\link[graphics]{plot.histogram}}, \code{\link[graphics]{plot.table}}
}
